Bug#24345509: WRONG RESULTS WHEN GCOL INDEX IS USED
When InnoDB evaluates an end-range condition while scanning a
non-covering secondary index, the value of any virtual generated
column referenced by the end-range condition is not available yet. The
evaluation of the end-range condition may therefore give an incorrect
result and cause the scan to stop before all qualifying records have
been fetched.

This is fixed by making the handler check if the pushed down end-range
condition references a virtual column, and set a flag if it does.
InnoDB checks the flag and, if necessary, reads the virtual column
values from the secondary index before the end-range condition is
evaluated.