Bug#23264552 -- XA: ASSERT `M_STATUS == DA_ERROR' IN MYSQL_ERRNO:SQL_ERROR.H:385
In case a storage engine failed to prepare a branch of an XA transaction
the method handlerton::prepare() returns nonzero value in the function ha_prepare().
To handle error condition the function ha_rollback_trans() is called to rollback
current transaction. While doing rollback the following sequence of functions is called
  ha_rollback_trans --> TC_LOG::rollback --> ha_rollback_low -->
    XID_STATE::set_error --> Diagnostics_area::mysql_errno.
It results in firing of the assert DBUG_ASSERT(m_status == DA_ERROR) since
no error has been set in Diagnostics_area. That leads to abnormal server
termination for the server built in DEBUG mode.

To fix it extra condition added in the function ha_rollback_low() to call
XID_STATE::set_error() only in case an error is set in Diagnostics_area.