Bug #24450908	UNDO LOG EXISTS AFTER SLOW SHUTDOWN
Problem:
========

1) cached undo segment is not removed from rollback segment history
(RSEG_HISTORY) during slow shutdown. In other words, If the segment is
not completely free, we are failing to remove an entry from the history
list. While starting the server, we traverse all rollback segment slots
history list and make it as list of undo logs to be purged in purge queue.
In that case, purge queue will never be empty after slow shutdown.

2) Freeing of undo log segment is linked with removing undo log header from history.

Fix:
====
1) Have separate logic of removing the undo log header from history list
from rollback segment slots and remove it from rollback segment history
even though it is not completely free.

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>
Reviewed-by: Marko Mäkelä <marko.makela@oracle.com>
RB:13672