Follow up patch for Bug#23048816:
Yet another attempth of fixing the failing ATR testcase 'testSpj -n bug#23048816'.
(Not able to reproduce in local sandbox)

The tescase use ERROR_INSERT(17531) to set 'ErrorSignalReceive= DBSPJ'.
That will cause future long memory section allocations to fail
with an 'out of memm' type of error. The intention for this
testcase was to trigger an 'out of long memory' error for the
*next* SPJ request delivered, such that the error recovery
in Dbspj::execSIGNAL_DROPPED_REP() should be tested.

However, as ErrorSignalReceive was set already at the start of the
preceeding REQuest, there could be other long memory allocations
requested before the next SPJ REQuest arrived - which also
would fail with an 'out of memory' error. That made the testcase
unpredictable, and we got failures not expected by the testcase.

This patch moves the ERROR_INSERT(17531) setting of
'ErrorSignalReceive= DBSPJ' into the code where the preceding
SPJ request has sent its final reply. No long memory allocations
from the SPJ block should then be expected until the next SPJ
REQuest arrives.