Bug#24661523: -DMYSQL_MAINTAINER_MODE=ON BUILD FAILURE ON MAC OS X,               8.0 XCODE TOOLCHAIN
This patch fixes a build problem with OS X 10.11 + XCode 8.
MacOS 10.12 adds support for the clock_gettime function
and this function is present in headers supplied with XCode 8.
Since we only tested if the function was declared, we got
linking errors with OS X 10.11 + XCode 8 since 10.11 does
not include the implementation of clock_gettime.

The problem is fixed by replacing the check with
CHECK_C_SOURCE_RUNS() CMake check which does linking as well.

The problem originally reported in the bug report is already
fixed in 8.0.1 by:
  Author: Abhishek Ranjan <abhishek.ar.ranjan@oracle.com>
  Date:   Thu Aug 25 19:43:10 2016 +0530

    WL#6392 : Upgrade to Transactional Data Dictionary

    Post push fix:

    Compilation with clang fails with error:
    unused variable 'event_table_def'