Bug#24670530	MTS WORKERS FAILS TO WAIT FOR COMPLETION OF DEPENDENT TRANSACTIONS
Problem: Dependent transaction check and conflict detection algorithm
is not working properly when the foreign key is referring to a unique key.

Analysis: In the hash key generation algorithm, we are prefix with letter
'P' for primary keys and we are prefix with 'U#' for unique keys.
And for foreign keys, we are generating the same key that would have got
generated for the referred key of the parent table. In the current code,
there was an assumption that foreign key always refers a primary key of
the parent table and hence it was prefixing 'P' while generating the key.
Those test scenarios that break the assumption i.e., foreign key is
referring unique key of the parent table, then the conflict detection
algorithm and dependent transaction index generation algorithm is not
working properly.

Fix: Instead of P and U# prefixes for the keys, we will prefix the key
names (index names). Same rule applies for the foreign key case as well.
The hash key for the foreign key will be prefixed with referred index
name.