Bug24569624: "SET PERSIST GENERAL_LOG=1" DOES NOT RESUME ON RESTART.
When a boolean system variable is persisted with "ON" as value the variable
is not pesisted with correct value. This is because in-correct Item_int class
is intantiated for the value "ON" and later when fix_fields() is called value
0 is saved, which results in wrong value.
Fix is to instantiate Item_string object for value "ON" or "OFF".