Bug#24352667: RELAY_LOG_BASENAME NULL POINTER CAUSE SLAVE CRASH
Problem:
========
Uninitialized 'relay_log_basename' will cause slave server
to exit abnormally.

Analysis:
=========
'relay_log_basename' variable is set only when 'relay_log'
option is specified in conf file. If the option is not
specified the 'relay_log_basename' is internally constructed
on the fly using 'hostname' but 'relay_log_basename'
variable is not set. When slave tries to access this
uninitialized variable it results in an abnormal exit of
server.

MySQL documentation states that the 'Default' value of
'relay_log_basename' is
"datadir + '/' + hostname + '-relay-bin'".
This means that in the absence of user specified relaylog
basename 'hostname' should be used to construct the basename
and the variable should be initialized with this value.

Fix:
===
At present code checks for the existence of user specified
option value to set 'relay_log_basename'. This check is not
required, in the case user has not specified this option
current hostname will be used to construct the
relay_log_basename.