Bug#23057045 CRASH WHEN RUNNING SHOW PROCESSLIST AFTER SYSBENCH RUN
Problem:
========
Srv_session only attaches/detaches to/from another session if the
new/old session is either not attached or it is attached to the current
thd. Since binlog attaches one by one to thd(s) by calling store_global,
but never restores them to previous state(restore_global) the check that
verifies if the owner of the physical thread is the same as the owner
of the THD object fails.

Solution:
=========
Made the destructor of Thd_backup_and_restore call restore_globals in
the destructor and also restored the previous real_id.

Change-Id: Icc14eb48a00b87dcef5b42cca633b8c6531011b7