BUG#24693798 PARTIALLY FAILED USER MANAGEMENT STMT FAILS TO CONSUME              GTID ON BINLOGLESS SLAVE
Problem:

The partially failed CREATE USER, RENAME USER and ALTER USER statements
does not consume a specified gtid when binary log is disabled.

Analysis:

All above mentioned statements can be partially failed when handling
multiple users, and a partially failed statement that is written to the
binary log when the binary logging is enable should consume a
specified gtid when GTID_NEXT is set and binary logging is not enabled.

Fix:

To fix these wrong behaviors, make a partially failed CREATE USER,
RENAME USER and ALTER USER to consume the specified gtid and save
the gtids into @@GLOBAL.GTID_EXECUTED (and mysql.gtid_executed table)
when binlog is disabled as it does when binlog is enabled.