Bug#24675602 LOCAL READ ON UNIQUE INDEX AND BLOB TABLES DO NOT WORK WITH TWO NODE GROUPS
For dirty and simple reads in fully replicated tables Dbtc requests local
fragment from Dbdih for a partition by calling DIGETNODESREQ with anyNode
set.

For reads in index and blob tables the dirty and simple flags are not
used in both key requests and scans, instead a read committed base flag
is used instead.

In two places use of the read committed base flag were added.