Bug#21271194 MISSING SANITY CHECK FOR MALLOC() IN SERVICES.CPP
A call to system function malloc is not checked for null in function
MgmApiSession::list_session. We cannot return error code for this failure
because as per prototype this function returns void. So, by default the
caller expects the function to be always success.

In this function, malloc is used to create a temp buffer, which we use to
create a message and write to the socket in single shot. If we
write the message to the socket directly we can avoid the temp buffer.

so, I fixed the bug by removing malloc and adding many write to the socket.