Bug#24577194 : LOAD DATA ERROR 1262 BEHAVIOUR DIFFERENT BETWEEN                5.6 AND 5.7 UNDER STRICT SQLMOD
WL#6891 reimplements STRICT mode in mysql-5.7.
In the scope of reimplementation, an error list is created which will
be affected by STRICT mode. ER_WARN_TOO_MANY_RECORDS was not added in
the list. This results in a regression from mysql-5.6.

As the counterpart error message ER_WARN_TOO_FEW_RECORDS is in the list
of the error to be affected by STRICT mode, the fix adds
ER_WARN_TOO_MANY_RECORDS to the list as well. This warning message is
only used by LOAD command.