Bug #24611754 XPROTOCOL: DON'T DROP THE CONNECTION IF YOU DON'T UNDERSTAND THE MESSAGE TYPE
Description:
A customer noticed that if he sends unknown message types to plugin
the current behaviour is to send an ERROR message of type HY000,
with text 'Invalid message' back to the client.
The customer also expects that the connection couldn't be dropped
but dedicated counter would be incremented.

Solution:
During the issue's investigation described dropping of connection
hasn't be confirmed. A dedicated counter for unknown messages
has been implemented as a status variable (Mysqlx_unknown_message_type).


Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>
RB:14112