Bug#24713918: ASSERTION `MDL_CHECKER::IS_READ_LOCKED(M_THD, *OBJECT)'               FAILED.
The problem was an assert during ALTER TABLE of a temporary
table if a normal table with the same name existed in the
same schema.

The reason was that we tried to check if the table to be
altered had any triggers. This meant accessing the data
dictionary which then found the normal table as temporary
tables are not part of the dictionary. We then asserted
since we did not have an metadata locks on the table name.
(Metadata locks are not taken for temporary tables).

This patch fixes the problem by not checking for triggers
when altering temporary tables as temporary tables cannot
have triggers.