Bug#23854015 OPTIMIZER FAILS TO LOOK UP RAND() IN SEMI-JOIN-TRANSFORMED SUBQUERY
SELECT_LEX::resolve_subquery(): Disable the semi-join transformation
if the left-hand-side of the IN predicate is nondeterministic.

The reason this problem occurs is that we try to create a ref access
on the subquery table using the result of a random function as
the lookup value. However, create_ref_for_key() is not designed to
process random values unless the ref access is created for
the last table in the join.

If there are nondeterministic components in the IN (subquery),
the subquery would not be subjected to semi-join transformation
at all. Instead, the left-hand-side of the IN (subquery) would be
transformed into a WHERE condition of the subquery.

Reviewed-by: Roy Lyseng <roy.lyseng@oracle.com>