Bug#24746530 PERFORMANCE_SCHEMA_MUTEX_INSTANCES_LOST=8 WHEN GROUP REPLICATION IS ENABLED.
Problem: When the group replication plugin is enabled, we are unable
to see few psi mutexes in the performance schema.setup_instruments
table.

Analysis: PFS_MAX_MUTEX_CLASS is declared as 200 and when the group
replication plugin is enabled, the number of psi mutexes is more
than this limit, hence we are loosing the mutexes. The same can be
viewed from Performance_schema_mutex_instances_lost variable.

Fix: Increase the limit to another safe limit , i.e., 220.