Bug#24732194: "USE DB_NAME" AND "SELECT SCHEMA()"               WORK FOR NON-EXISTING DATABASES
The problem was that USE checked for file directory existence,
rather than asking the data dictionary if the schema exists.

This mean that creating the schema manually (e.g. mkdir)
would first appear to work when using USE, but would then
later fail since the schema was not reflected in the dictionary.

This patch fixes the problem by changing USE to ask the
dictionary rather than the file system.

Note that creating schemas by creating directories manually
is not supported.