Bug #24716756   IMPROVE SCALABILITY OF API RECEIVER HANDLING
In order for the NDB API to process incoming signal signals, the
threads which wants to act as receiver has to acquire the 'poll-right'
from the transporter layer. This poll right could either be
requested and assigned to a separate receiver thread, or each client
thread can take the receiver role when it anyway had to wait for
its result.

When the thread being the poll owner had received a sufficient amount
of data, it will release locks on other clients which it has taken
while delivering signals to them. This may make them runnable again,
and the OS-scheduler might decide that it is time to wake them up.
This might happen at the expense of the poll owner threads, which
then is yielded from the CPU - while still holding the 'poll-right' !

This patch will release the poll-right from a thread
before unlocking and signaling other threads. Thus the poll right
becomes available for other threads that are actively
executing on the CPU.

This increase the concurrency in polling data of the receiver
and should also reduce latency for clients wating for being woken
up.