Bug#22512899 - DROP TRIGGER AFTER RENAME TABLE RESULTS IN ERROR CODE 1146: TABLE DOESN'T EXIST
In case a table has an associated trigger and the table is renamed to a new name
contained characters in both lowercase and uppercase then successive execution of
DROP TRIGGER for the table’s trigger would result in issuing the error
ER_NO_SUCH_TABLE for the table associated with the trigger.
This bug happens only for Mac OS X platform.

The reason for the bug was that case of a file name characters wasn’t ignored
when searching a file entry in the file system.

To fix the bug convert a table name to lowercase when resolve a trigger’s table
during handling the statement DROP TRIGGER.