Follow up of bug#24496910, intended to fix bug#24715897: Showstopper for 7.5 GA
The patch for bug#24496910 also introduced locking of the injector_mutex
when reading global variabled maintained by the binlog injector thread, and
read by any client thread.

As it turns out that the injector_mutex is held for intervals of 10ms,
that could considderably slow down operations as open'ing of tables
to a non acceptable level.

This locking of this mutex is reverted ndb_binlog_is_read_only().
A big comment is added explaining why, and suggesting TODO actions
for a later propper fix.