BUG#24595992: NULL POINTER DEREFERENCE IN CHECK_DUPLICATE_KEY()
Analysis:
========
CREATE TABLE using SELECT can cause a null pointer dereference
while printing the warning message.

During the processing of SELECT in CREATE TABLE, the first
element of the global query table_list is pruned. Later
while trying to print a warning message, it tries to access
the global query table list to get the schema name and
table name which is set to NULL.

Fix:
===
Passed the schema name and table name from the upper layer
so that it is accessible in check_duplicate_key().