Bug #24788778: SEVERE REGRESSION IN MY_STRNXFRM() FROM MYSQL-5.5 -> 5.6
MySQL 5.6 changed utf8 binary collation strxfrm() from just comparing the UTF-8
string to converting it to UCS-2 (big-endian) and padding before compare.
(Similarly for utf8mb4, just with UCS-3 instead, which isn't correct for
all Unicode code points.)

Both are correct (although changing it also changed the hash function, which
might on-disk binary compatibility wrt. partitioning), but the latter makes for
fixed-length keys, which was seemingly important for the (now discontinued)
Falcon storage engine at the time. However, it also introduces a bottleneck
when hashing. We add a very simple benchmark (based on the benchmark in the
bug) and then optimize the string transformation:

 - Add an SSE2 version of the padding as long as we're sufficiently far away
   from the end of the string; this will be used automatically for all 64-bit
   Intel compiles (as well as 32-bit Intel compiles with -march=native or
   similar).
 - Inline the mbwc() function for the special case of my_utf8_uni.
 - Some general microoptimization.

All in all, we're at about 7x the speed of before. Old version
(trunk, 64-bit, optimized mode, GCC 6.1.1, Skylake 3.4 GHz):

  Done, used 1.696 seconds (1.696 us/iteration)

This version:

  Done, used 0.242 seconds (0.242 us/iteration)

The code is written to be C++98 compatible (even though a lambda would be
much more elegant than a functor class), as it will probably want a 5.6 and 5.7
backport. The benchmark isn't, though, as it uses std::chrono.

Change-Id: I9da8c8af448bdacd6028a65c10ce232f841cdc94