Patch for bug#24800098 NDB: OPTIMIZE NDBSQLUTIL::STRNXFRM_BUG7284()
As the optional space padding of the transformed string is not
required for a well behaved character collation (>99%), we
should hardly never need to produce a space-pad string.

This patch change ::strnxfrm_bug7284() such that this string
is only produced when actually needed. In most other cases the
result produced from the string library method strnxfrm(,,, PAD_)
is returned 'as is'.