Bug #23227428: SQL PLAN IS NOT ACCORDING WITH                OPTIMIZER_TRACE
Issue:
------
While choosing an index for a query with ORDER BY and
LIMIT, optimizer's switch to a different index is not
recorded by the trace.

Solution:
---------
This is a side-effect of the fix for Bug#18035906. While
tracking changes to the preferred index, no distinction is
made between index scan and ref access. This results in
misinterpretation of whether a new index has been chosen.