Bug#24813358: AUTO_RELEASER SHOULD OWN OBJECTS ACQUIRED BY ACQUIRED_UNCACHED
After this refactoring, dictionary objects acquired by
Dictionary_client::acquire_uncached() are owned by the current
auto releaser and not by the caller. This means that deletion
of such object will happen when the auto releaser goes out of
scope instead of being up to the caller.

This patch also changes acquisition of view objects in TABLE_SHARE
from acquire_uncached() to acquire() + clone(). This is safe since
we have a metadata lock on the view at this point. The advantage
is that we now can use the dictionary cache and thus avoid
accessing disk.

Finally, the patch fixes a latent MTR bug. In check-testcase, the
I_S query used to retrieve routine info did not have an ORDER BY
clause, making the order of objects returned unstable.