Bug #24482919 -- SHOW TRIGGER OUTPUT INCONSISTENCY ON WINDOWS / LINUX ON 8.0.0
After WL#7896/WL#6383 an output of SHOW TRIGGERS could produce inconsistent
result on Linux and Windows platform.

The reason of the bug was the wrong compare condition that was used to sort
trigger objects loaded from the Data Dictionary into the memory.

To fix the bug the sorting criterion has been modified to make it in
lexicographical order.