Bug#24814874	FAILING TEST IN PB2 DAILY-5.7 INNODB.INNODB-MULTIPLE-TABLESPACES
Problem
-----------
The test is testing the sceanrio where valid ibd files are present
in different valid paths during crash recovery.In case a
checkpoint happens just before we kill the server ,it will not go
into crash recovery,because of which the test case was failing.

Fix
---
We have used helper functions to detect if checkpoint happened before the
intended crash of the server.If yes we skip the test.

[Approved by satya over IM ]