Bug #24801602ADD CHECK_PLUGIN_DIR.INC TO VERIFY PLUGINS LOCATION IN THE TEST.
In mysql source, all the plugins might not be in the same location. Some mtr
tests would require multiple plugins, so it is desirable that there is a
check to see that, if all the plugins are in a directory, then the test
should run, else it needs to be skipped.

Steps involved in this inc file :
1. Take a plugins list.
2. check if all plugins are in same directory.
3. If yes, continue with the test.
4. If no, Skip the test.