Bug#24826541 TABLE PERFORMANCE_SCHEMA.DATA_LOCKS IS MISSING RECORDS
Tests involving performance_schema.lock_data have been found
to be unstable.
After analysis, several issues, both in the server code and the test
scripts, have been identified, fixed by this patch.

ISSUE #1
========

When SELECT * from performance_schema.data_locks
is executed, the code in table_data_locks::rnd_next()
extracts rows from innodb, using a restartable scan,
using an iterator:
  PSI_engine_data_lock_iterator *it= ...

Assume that for example the first call to
  it->scan();
returns 10 rows, and a subsequent call returns 5 more rows.

Before this fix,
- the data container after first call contains rows [0] .. [9]
- the ::rnd_next loop with m_index_2 at positions [0] .. [9]
- all the rows for the first scan are returned properly
- the data container after second call contains rows [0] .. [5]
- the ::rnd_next loop with m_index_2 at positions [10] .. [14]
- rows for the second scan are missed

The root cause is the discrepancy in "index":
- the data container uses relative positions in each scan
- the m_index_2 loop in rnd_next uses absolute positions

With this fix, restartable scans keep rows numbered in absolute,
with PFS_data_lock_container::m_logical_row_index.
The container is not cleared, but only truncated with ::shrink(),
during the scan, to preserve row numbering.

ISSUE #2
========

Because of restartable scans, locks are not returned always in the same
order from a select, as two transactions can be returned
in the same or in different scans.

As a result, tests scripts need to use an ORDER BY clause
on SELECT, to have a stable output.

This fix corrected the test scripts.

ISSUE #3
========

Using SELECT ... ORDER BY cause the optimizer to use
table_data_locks::rnd_pos(), after a filesort.

This code path is longer from a plain table scan,
and exposed further bugs:

ISSUE #3-1
----------

Lock rows exported by a scan can not be fetched back,
because of incorrect filtering on the lock heap_no.

A given LOCK_REC structure is not associated with a
single heap_no, but with a -- set of -- heap_no.

When performing a table scan,
Innodb_data_lock_iterator::scan_trx()
is called with filter = false,
and iterate properly on heap_no,
using:
  heap_no = lock_rec_find_set_bit(lock);
  heap_no = lock_rec_find_next_set_bit(lock, heap_no);

When called for a fetch however, with filter = true,
the code failed to look for the entire set of heap_no,
and only considered the first in the set.

This fix now iterate properly on all heap_no
when using a filter with filter_heap_id.

ISSUE #3-2
----------

Transactions exported by a scan can not be fetched back,
because of incorrect iteration on the trx list in innodb.

The "transaction list" in innodb is a complex structure,
made of two lists (trx_list, mysql_trx_list).

To iterate properly, the code must use the get_next_trx()
helper, instead of using UT_LIST_GET_NEXT().

The code in fetch_trx_in_trx_list() has been fixed to iterate
on the proper transaction list.

ISSUE #3-3
----------

Transactions exported by a scan can not be fetched back,
because of incorrect values for the transaction id.

The problem is that trx->id may, or may not,
be a valid transaction id.

In particular, some transactions in read only mode
can have a trx->id of 0.

See comments in trx_get_id_for_print() for details.

The code exported a lock with a value for engine_lock_id,
that contained a (fabricated) transaction id value.

On a subsequent fetch, after parsing the value of engine_transaction_id
to find a trx id, fetch_trx_in_trx_list() failed to locate the transaction,
because the code was comparing the fabricated value with a zero trx->id

The fix is to always use trx_get_id_for_print(trx), and never use trx->id
directly, when exporting data in a scan, or fetching a particular row from
rnd_pos().

ISSUE #4
--------

In the test scripts, replying on column LOCK_DATA
in table performance_schema.data_locks may cause spurious failures.
This column can be NULL, in case the innodb storage engine
purged data from the server memory to disk: this is a documented limitation.

Relaxed the use of LOCK_DATA in test cases.