Bug#22924183 GROUP BY ON DERIVED TABLE, EXPRESSION OF AGGREGATE AND GROUP COLUMN: ERROR
When a query was doing GROUP BY on a column of a derived column,
and the SELECT list contained an expression mixing an aggregate function
and the group column, an error was raised if only_full_group_by
was enabled.

Cause:
consider:
1) SELECT (je+1)+count(*) FROM (SELECT doc+1 AS je FROM coll) AS dt GROUP BY je;
2) SELECT je+(1+count(*)) FROM (SELECT doc+1 AS je FROM coll) AS dt GROUP BY je;

an Item_direct_view_ref is produced for 'je' in the select list.
Then in Item::split_sum_func2() an Item_aggregate_ref is produced for
items involved with count(*);
when it's for (je+1) (first query), this Item_aggregate_ref wraps '+' which
itself has one argument pointing to Item_direct_view_ref.
only_full_group_by logic looks at the selected expression;
it considers Item_aggregate_ref to be a "transparent"
item so it looks through it and finds Item_direct_view_ref, which is equal
to the group column so the query is accepted.
When it's for je+() (second query), Item_aggregate_ref _replaces_
Item_direct_view_ref, due to this:
    evgen@moonbone.local <>  2005-09-07 20:38:36
    Fix bug # 12922 if(sum(),...) with group from view returns wrong results
So Item_aggregate_ref wraps the expression defining derived column (doc+1),
which isn't the group column ('je'), hence the error.

Fix:
do not replace Item_direct_view_ref, do not wrap the real_item() inside Item_aggregate_ref;
instead, wrap Item_direct_view_ref inside Item_aggregate_ref; it's more logical as
code already works this way in the case where Item_direct_view_ref is deeper
nested (example: when Item_direct_view_ref is argument of an additional function;
indeed a workaround to this bug was to replace je+(1+count(*)) with
coalesce(je)+(1+count(*)), which, by adding an extra layer, prevented the
removal of Item_direct_view_ref, making Item_aggregate_ref simply wrap coalesce()).