Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
Prefactoring: Clean up the mb_wc functions for UTF-8 a bit.

my_utf8_uni is renamed to my_mb_wc_utf8 to be consistent with my_mb_wc_utf8mb4.
They are both made into inline functions in a header file, so that they can be
used inline in ctype-uca.cc (this will be important in a future patch in the series).
The mb4 variant is given a functor like the we already had with the mb3 variant
(again, this will be important in a future patch), and gets its thunks.

Finally, the mb4 variant now drops its unused cs header.

Performance for functions like my_hash_sort_utf8mb4 probably goes up a bit
due to the inlining, but I haven't tried measuring (it's probably too small
to be seen on sysbench, and we have no microbenchmaks for them).

Change-Id: Iedf2b4cf894a2badcc4d832215e5a72654df3c26