Bug#24489037 ASSERTION `!\"SHOULD NOT HIT HERE\"\' IN DD_GET_OLD_FIELD_TYPE
The issue here is that, some of UDF's introduced by WL#6599
crashes because the UDF is invoked with wrong argument i.e.,
NULL. These UDF's are invoked with data dictionary table
columns as arguments. And the UDF's mostly expected valid
arguments and did not check for NULL. The crash is due to a
assert when we got a invalid column type. This can happen if
a user has a query which does a LEFT JOIN and I_S table.
Then there is possiblity that these UDF's getting called
with NULL as their arguments.

Note that these UDF's are used by INFORMATION_SCHEMA system
views and are not exposed to users.

The patch also does following changes:-

- Makes sure all the UDF's do proper evaluation of
  arguments.

- Removed Item_*::fix_length_and_dec() methods which are no
  more used by optimizer. Replaced it with adding
  resolve_type() method as requested by Optimizer team.

- Pass only table 'options' to UDF INTERNAL_KEYS_DISABLED,
  as schema and table name are not really used.

- Few UDF's are marked as nullable, and others are retained
  as non-nullable to match the old behavior expected.

- A test case invoking all UDF's with NULL arguments is
  added to main.dd_is_gcov.