Bug#25047032 DECLARE OVERRIDDEN ITEM MEMBER FUNCTIONS AND FINAL ITEM SUBCLASSES
When virtual member functions are being renamed or replaced, it would
be extremely useful if the compiler were able to warn about member
functions that were intended to be overriding a member function in a
base class, but are not actually overriding anything.

The C++11 override keyword would allow this.

Furthermore, the C++11 final keyword would enable some additional
optimizations, and it would make the code more readable by documenting
that there cannot be any classes derived from a final class.

This is a logical follow-up of
Bug#24787984 DECLARE DERIVED METHODS IN THE OPTIMIZER WITH THE C++11
OVERRIDE KEYWORD
applied to all Item classes.

Reviewed-by: Knut Hatlen <knut.hatlen@oracle.com>