Bug#24519021 FEW XA TESTS FAIL IN WAIT_COND'N.INC AS VIEW ON P_S.THREADS DOESNT RETURN RESULT
This is a regression from WL#6616 Performance Schema, Indexes. Under high
stress conditions, performance_schema.threads may attempt to materialize a
row representing a thread that is being disconnected, resulting in an error
similar to "At line 41: Query 'SELECT count(*) = 0 FROM v_processlist WHERE
PROCESSLIST_ID = 17' didn't return a result set

The fix is for table_threads::make_row() to return a row only if the state of
the thread has not changed.