Bug #24823885: PERFORMANCE REGRESSION WHEN CHANGING CHARACTER SET TO UTF8MB4
Specialize the UCA scanners on number of levels to compare.
In particular, this allows us to make the scanner significantly
simpler for the special case of only one level (ie., accent-
and case-insensitive collations), making for a significant
speed boost.

Also add a new accent- and case-_sensitive_ benchmark, so that
we can track the difference in speedup between the two.

Microbenchmarks (Skylake, 3.4 GHz, 64-bit, GCC 6.1.1):

  BM_SimpleUTF8MB4       1374 -> 1138 ns/iter  (+20.7%)
  BM_MixedUTF8MB4         597 ->  527 ns/iter  (+13.3%)
  BM_MixedUTF8MB4_AS_CS  1839 -> 1760 ns/iter  ( +4.5%)

sysbench results are pretty much in the noise (+3–4% for a
single-threaded test).

Change-Id: I942d178962b053e2f65dbb7de1f4ffc61a98ea81